refactor: Put tests in same module that defines struct
